Fix extracting references from regex match
The regex uses separate capture groups for alternative expressions that
match a single-quoted or a double-quoted string. This is more robust
with strings that may contain quotes.

However the code using the match objects doesn't currently take the two
capture groups into account and always expects the results in group 1,
which only matches double-quoted strings.

This change introduces a check so that if group 1 is None, the function
will use the matched substring in group 2.

Closes #36